images get loaded without being hardcoded
1:26

as of now, its in the testing side of things because i dont want to mess with the main one untill I have EVERYTHING figured out. i had to make a few python scripts for myself to use in order to log the images from figma. i have MUCH more to do before it can be deployed but i forgot what it is exacally... ill do images then think of what needs to get impliminted