Finishing path 1
2:38

I made sure the user has a way to go back to the choice where they messed up whenever they died, i made a counter of death that gets displayed on the end page, also everything is properly linked. I also added the achievements. there are 3 right now in the first path. i need to figure out how to get that'crappy' video in and have it overide whatever i have in the code rn