png -> webp
changed the images to webp (the origionals are still in there), then added some more js to forcedpath.js for auto loading the images... lets see if it works.

(also  i changed the png endings to webp so if you;re testing with a png, might want to change all the webp's to that. it shouldent take very long)