code for icon
alright there's still a little bit more work i have to do so first is the proper icon img for each page, i have the code up and running. next is the code to get rid of the fade in/ fade out because i think its messing with the images loading.